requirements.txt
psycopg2
django>=1.11,<1.12
djangorestframework>=3,<4
pillow
pytest
pytest-django
pytest-cov
freezegun
flake8

